add genuxMiddleware inside authenticationMiddlleware.js
