+++ src/middlewares/authenticationMiddleware.js
@@ -1,6 +1,6 @@
 const got = require('got');
 
-module.exports = function authenticationMiddleware() {
+const authenticationMiddleware = () => {
   const authServerClient = got.extend({
     prefixUrl: process.env.AUTH_SERVER_URL
   });
@@ -11,7 +11,7 @@ module.exports = function authenticationMiddleware() {
       return res.status(400).json({ reason: 'Missing access token' });
     }
 
-    authServerClient.post('validateaccesstoken', { 
+    authServerClient.post('validateaccesstoken', {
       json: { accessToken: token }
     })
       .then(result => next())
@@ -20,4 +20,26 @@ module.exports = function authenticationMiddleware() {
         res.status(401).json(err.response.body);
       });
   }
+};
+
+const genuxMiddleware = () => {
+  const authServerClient = got.extend({
+    prefixUrl: process.env.AUTH_SERVER_URL
+  });
+
+  return (req, res, next) => {
+    authServerClient.post('useGenuxToken', {
+      json: { genuxToken: req.body.genuxToken }
+    })
+      .then(result => next())
+      .catch(err => {
+        console.error(err.response.statusCode, err.response.body);
+        res.status(err.response.statusCode).json(err.response.body);
+      });
+  }
+};
+
+module.exports = {
+  authenticationMiddleware,
+  genuxMiddleware
 };
