+++ src/routes/establishmentsRouter.js
@@ -1,6 +1,6 @@
 const express = require('express');
 
-const authenticationMiddleware = require('../middlewares/authenticationMiddleware')();
+const { authenticationMiddleware } = require('../middlewares/authenticationMiddleware');
 
 const visitManagerGateway = require('../gateways/visitManagerGateway');
 const establishmentsController = require('../controllers/establishmentsController')(visitManagerGateway());
@@ -9,11 +9,11 @@ module.exports = function establishmentsRouter() {
   return express.Router().use(
     '/establishments',
     express.Router()
-      .get('/', authenticationMiddleware, establishmentsController.get)
+      .get('/', authenticationMiddleware(), establishmentsController.get)
       .post('/', establishmentsController.add)
-      .get('/:establishmentId', authenticationMiddleware, establishmentsController.getSingleEstablishment)
-      .get('/PDF/:establishmentId', authenticationMiddleware, establishmentsController.getEstablishmentPDF)
-      .put('/:establishmentId', authenticationMiddleware, establishmentsController.update)
-      .delete('/:establishmentId', authenticationMiddleware, establishmentsController.remove)
+      .get('/:establishmentId', authenticationMiddleware(), establishmentsController.getSingleEstablishment)
+      .get('/PDF/:establishmentId', authenticationMiddleware(), establishmentsController.getEstablishmentPDF)
+      .put('/:establishmentId', authenticationMiddleware(), establishmentsController.update)
+      .delete('/:establishmentId', authenticationMiddleware(), establishmentsController.remove)
   );
 };
